implement in-memory persistence of crawled pokemons
This introduces the Grid, aimed to store the pokémons (and maybe other
geo-localized informations). The Grid groups object in cells of 100x100
meters. This feature isn't used for now but should be used to have a
more intelligent crawler displacement (ex: have a "last update date" on
each grid cell).

Also, use iterators because it's fun.