implement a smarter crawling strategy
Every time a user scans a location, it is registered as a "point of
interest" (POI) for 10 minutes. POIs are alternatively scanned so each
user can have his/her own interesting part of the map updated in a fair
manner. The user POI is renewed every 5 minutes if he/she is still using
the map, thus no crawling is done when no user is using the map.